requirements2.txt
acme==1.21.0
attrs==21.2.0
Automat==20.2.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
CacheControl==0.13.1
cached-property==1.5.2
cachetools==5.3.2
certbot==1.21.0
certbot-nginx==1.21.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
ConfigArgParse==1.5.3
configobj==5.0.6
constantly==15.1.0
cryptography==37.0.0
dbfread==2.0.7
decorator==4.4.2
defer==1.0.4
distro==1.7.0
docopt==0.6.2
docutils==0.17
duplicity==1.0.0
ebaysdk==2.1.5
fasteners==0.14.1
firebase-admin==6.3.0
freezegun==1.1.0
future==0.18.2
geoip2==2.9.0
gevent
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.13.0
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==1.1.2
grpcio==1.60.0
grpcio-status==1.48.0
httplib2==0.20.2
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.6.4
incremental==21.3.0
isodate==0.6.1
jeepney==0.7.1
Jinja2==3.0.3
josepy==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libsass==0.20.1
lockfile==0.12.2
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
maxminddb==2.5.2
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.7
netifaces==0.11.0
num2words==0.5.10
oauthlib==3.2.0
ofxparse==0.21
olefile==0.46
paramiko==2.9.3
parsedatetime==2.6
passlib==1.7.4
pdfminer.six==20231228
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.1.0
ply==3.11
polib==1.1.1
proto-plus==1.23.0
protobuf==3.21.12
psutil==5.9.0
psycopg2==2.9.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.20.0
pydot==1.4.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyICU==2.8.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.1
python-ldap==3.4.0
python-magic==0.4.24
python-stdnum==1.17
pytz==2022.1
pyusb==1.2.1
pyxdg==0.27
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==1.0.0
rjsmin==1.1.0
rsa==4.9
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
soupsieve==2.5
ssh-import-id==5.11
Twisted==22.1.0
uritemplate==4.1.1
urllib3==1.26.5
vobject==0.9.6.1
wadllib==1.3.6
Werkzeug==2.0.2
xdg==5
xlrd==2.0.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.1.0
zipp==1.0.0
zope.component==4.3.0
zope.event==5.0
zope.hookable==5.1.0
zope.interface==5.4.0

